Implement lazy-loading and UI cleanups
Introduce React.lazy + Suspense for page-level code splitting (with a Loader2 fallback) and add a /login placeholder. Refactor AppLayout and AppSidebar (clarifying comments), add utility docs (cn helper) and multiple small UI/markup cleanups across pages (AlertaUrgente, Alertas, CampanhaDetalhes, Campanhas, Configuracoes, Index): compact JSX, adjust form/dialog markup, minor text tweaks, remove an unused delete button, and add team-facing comments. Also improve data loading in Index to use Promise.all and ensure Alertas marks-as-read updates state immediately. Commit includes numerous backend build artifacts and newly added/updated API controller/domain/repository/service classes (e.g. Noticia, OrgaoSaude, UsuarioService) generated by the build.

+++ conecta-a-vida-api/src/main/java/br/edu/ifsp/conectaavida/domain/Usuario.java
@@ -1,31 +1,22 @@
 package br.edu.ifsp.conectaavida.domain;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Entity
 @Table(name = "usuarios")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
+@Getter @Setter @NoArgsConstructor @AllArgsConstructor
 public class Usuario {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    // ATENÇÃO EQUIPE: ID como Integer por causa da tabela existente no Supabase.
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
-    @Column(nullable = false)
-    private String nome;
+    @Column(nullable = false) private String nome;
 
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String senha;
+    // unique = true avisa ao banco para não aceitar dois usuários com o mesmo e-mail.
+    @Column(nullable = false, unique = true) private String email;
+    @Column(nullable = false) private String senha;
 
     private Integer idade;
     private String sexo;
